requirements.txt
numpy
anthropic
Pillow
pillow-heif
pydantic
python-dotenv
sqlalchemy
sqlmodel
mlb-statsapi
